+++ src/global.d.ts
@@ -1,5 +1,3 @@
 // import sureUtils from "./index";
-
 // declare namespace sureUtils {
-
 // }